requirements.txt
# === 2025 Mid-Year Stable RAG Stack Requirements ===
# This file defines the TOP-LEVEL dependencies for the project.
# Pip will automatically handle the installation of their sub-dependencies.

# --- Web Application Framework ---
# For creating the user interface and web server.
flask==3.0.3
python-dotenv==1.0.1

# --- Core AI & LangChain Framework ---
# The main framework for building the RAG chain.
langchain==0.2.5
langchain-core==0.2.9
langchain-community==0.2.4

# --- Hugging Face Ecosystem ---
# For loading models from Hugging Face Hub.
langchain-huggingface==0.0.3  # Bridge between LangChain and HF
huggingface-hub==0.23.0       # Client for downloading models and interacting with the Hub
sentence-transformers==2.7.0  # For running the embedding model
transformers==4.41.2          # Core library by Hugging Face, a dependency for the above

# --- Machine Learning Backend ---
# The foundational library for all model operations.
torch==2.3.1

# --- Vector Store & Document Processing ---
# The local vector database.
faiss-cpu==1.8.0
# For loading and parsing PDF files.
pypdf==4.2.0

# --- Essential Utilities ---
# A core dependency for numerical operations in many libraries.
numpy==1.26.4
# For safe HTML rendering in Flask templates.
markupsafe==2.1.5